Even better merge/reconcile code
Drop the whole LocalObj thing as it was specific to the incr reconcile
fun and it can be done more simply by summing the max of each node
vclock counter.  Yes, I am exploting the impl details of vclocks here
and it probably isn't idiomatic, but it serves as an example.